+++ src/buildings/interfaces/Building.java
@@ -4,7 +4,7 @@ public interface Building
 {
     int floorCount(); // кол-во этажей в здании
     int spaceCount(); // общее число помещений в здании
-    int totalSpaceArea();   // общая площадь всех помещений здания
+    double totalSpaceArea();   // общая площадь всех помещений здания
     int totalRoomsCount();  // кол-во всех комнат зданич
     Floor[] getFloorArray(); // получить массив этажей здания
     Floor getFloor(int index);
